Add option to use a trait instead of extending a model class
